backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0

langgraph==0.2.6
langchain==0.2.12
langchain-aws==0.1.17
langchain-community==0.2.12
langchain-postgres==0.0.10

psycopg2-binary==2.9.9
pgvector==0.3.2

boto3==1.34.162

openai-whisper==20250625

TTS==0.22.0

numpy==2.4.3

torch==2.11.0
torchaudio==2.3.1

soundfile==0.12.1
pydub==0.25.1

python-dotenv==1.0.1

pydantic==2.8.2
pydantic-settings==2.4.0

sqlalchemy==2.0.32

aiofiles==23.2.1
httpx==0.27.0